# Optimize the optimization code