--- a/geometry/symmetric_bilinear_form.hpp
+++ b/geometry/symmetric_bilinear_form.hpp
@@ -336,6 +336,7 @@ using internal_symmetric_bilinear_form::Anticommutator;
 using internal_symmetric_bilinear_form::InnerProductForm;
 using internal_symmetric_bilinear_form::SymmetricBilinearForm;
 using internal_symmetric_bilinear_form::SymmetricProduct;
+using internal_symmetric_bilinear_form::SymmetricSquare;
 
 }  // namespace geometry
 }  // namespace principia