Note how the tests show that we call the functions less often.

Before:
```
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
BM_MLSLBranin/10/10             187858 ns       187752 ns         3739 average minima: 3.22011
BM_MLSLBranin/20/10             389219 ns       391089 ns         1795 average minima: 3.24513
BM_MLSLBranin/50/10            1081732 ns      1095171 ns          641 average minima: 3.30577
BM_MLSLBranin/10/20             393715 ns       379608 ns         1726 average minima: 3.28911
BM_MLSLBranin/20/20             843044 ns       852179 ns          897 average minima: 3.31438
BM_MLSLBranin/50/20            2397377 ns      2400015 ns          299 average minima: 3.34448
BM_MLSLBranin/10/50            1102934 ns      1095171 ns          641 average minima: 3.39782
BM_MLSLBranin/20/50            2500990 ns      2451444 ns          280 average minima: 3.4
BM_MLSLBranin/50/50            7353703 ns      7382190 ns          112 average minima: 3.41071
BM_MLSLGoldsteinPrice/10/10     165868 ns       166883 ns         4487 average minima: 2.83151
BM_MLSLGoldsteinPrice/20/10     348035 ns       344120 ns         2040 average minima: 2.93971
BM_MLSLGoldsteinPrice/50/10     982204 ns       973485 ns          641 average minima: 3.55382
BM_MLSLGoldsteinPrice/10/20     354820 ns       359414 ns         2040 average minima: 3.11275
BM_MLSLGoldsteinPrice/20/20     784468 ns       771663 ns          748 average minima: 3.38904
BM_MLSLGoldsteinPrice/50/20    2222311 ns      2235528 ns          321 average minima: 3.96262
BM_MLSLGoldsteinPrice/10/50    1024695 ns      1022159 ns          641 average minima: 3.69891
BM_MLSLGoldsteinPrice/20/50    2429662 ns      2452190 ns          299 average minima: 3.97993
BM_MLSLGoldsteinPrice/50/50    6940219 ns      6864044 ns          100 average minima: 4
BM_MLSLHartmann3/10/10          395216 ns       391089 ns         1795 average minima: 2.87465
BM_MLSLHartmann3/20/10          747345 ns       747831 ns          897 average minima: 2.96767
BM_MLSLHartmann3/50/10         1999140 ns      1989578 ns          345 average minima: 3.0087
BM_MLSLHartmann3/10/20          782582 ns       782614 ns          897 average minima: 2.98328
BM_MLSLHartmann3/20/20         1655374 ns      1625662 ns          499 average minima: 3.00401
BM_MLSLHartmann3/50/20         4710297 ns      4784031 ns          150 average minima: 3.01333
BM_MLSLHartmann3/10/50         2106231 ns      2089733 ns          321 average minima: 3.01558
BM_MLSLHartmann3/20/50         4755549 ns      4733823 ns          145 average minima: 3.01379
BM_MLSLHartmann3/50/50        14870207 ns     14213424 ns           45 average minima: 3
```
After:
```
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
BM_MLSLBranin/10/10             149747 ns       149500 ns         4487 number of minima: 3.22041
BM_MLSLBranin/20/10             302485 ns       298932 ns         2244 number of minima: 3.24955
BM_MLSLBranin/50/10             825540 ns       834788 ns          897 number of minima: 3.29543
BM_MLSLBranin/10/20             307436 ns       305884 ns         2244 number of minima: 3.29545
BM_MLSLBranin/20/20             656580 ns       667384 ns         1122 number of minima: 3.31105
BM_MLSLBranin/50/20            1920282 ns      1918729 ns          374 number of minima: 3.35829
BM_MLSLBranin/10/50             856055 ns       869571 ns          897 number of minima: 3.41026
BM_MLSLBranin/20/50            1939263 ns      1944360 ns          345 number of minima: 3.4058
BM_MLSLBranin/50/50            5798614 ns      5850038 ns          112 number of minima: 3.41071
BM_MLSLGoldsteinPrice/10/10     132866 ns       133501 ns         5609 number of minima: 2.83687
BM_MLSLGoldsteinPrice/20/10     272587 ns       269075 ns         2493 number of minima: 2.94224
BM_MLSLGoldsteinPrice/50/10     765998 ns       765222 ns          897 number of minima: 3.54627
BM_MLSLGoldsteinPrice/10/20     278860 ns       275333 ns         2493 number of minima: 3.12074
BM_MLSLGoldsteinPrice/20/20     604700 ns       583961 ns         1122 number of minima: 3.3984
BM_MLSLGoldsteinPrice/50/20    1742716 ns      1758835 ns          408 number of minima: 3.96814
BM_MLSLGoldsteinPrice/10/50     852182 ns       834230 ns          748 number of minima: 3.69786
BM_MLSLGoldsteinPrice/20/50    1752047 ns      1758835 ns          408 number of minima: 3.98529
BM_MLSLGoldsteinPrice/50/50    5592370 ns      5571464 ns          112 number of minima: 4
BM_MLSLHartmann3/10/10          308739 ns       312836 ns         2244 number of minima: 2.87299
BM_MLSLHartmann3/20/10          579272 ns       570057 ns         1122 number of minima: 2.96524
BM_MLSLHartmann3/50/10         1563568 ns      1563484 ns          449 number of minima: 3.00445
BM_MLSLHartmann3/10/20          603058 ns       597865 ns         1122 number of minima: 2.98217
BM_MLSLHartmann3/20/20         1240578 ns      1251345 ns          561 number of minima: 3.00178
BM_MLSLHartmann3/50/20         3577807 ns      3600023 ns          195 number of minima: 3.01026
BM_MLSLHartmann3/10/50         1636310 ns      1632973 ns          449 number of minima: 3.01336
BM_MLSLHartmann3/20/50         3748739 ns      3754035 ns          187 number of minima: 3.0107
BM_MLSLHartmann3/50/50        12017576 ns     11978648 ns           56 number of minima: 3
```
#3358.